# Use math/bits.Reverse32 instead of bitReverse