--- a/go.mod
+++ b/go.mod
@@ -1,0 +1,3 @@
+module github.com/jfreymuth/vorbis
+
+go 1.15