Merge #7 first.

The performance doesn't change on MacBook Pro 2020.

```
name     old time/op  new time/op  delta
IMDCT-8  54.3µs ± 0%  53.8µs ± 0%   ~     (p=0.100 n=3+3)
```